requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.6
django-ckeditor==6.5.1
django-cors-headers==3.13.0
django-environ==0.9.0
django-js-asset==2.0.0
django-richtextfield==1.6.1
django-timestampable-model==0.1.4
djangorestframework==3.14.0
mysqlclient==2.1.1
Pillow==9.4.0
psycopg2-binary==2.9.5
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
